pyproject.toml
[project]
name = "chat-app-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "attrs==25.1.0",
    "autobahn==24.4.2",
    "automat==24.8.1",
    "autopep8==2.3.2",
    "cffi==1.17.1",
    "channels==4.2.0",
    "channels-redis==4.2.1",
    "constantly==23.10.4",
    "cryptography==44.0.1",
    "daphne==4.1.2",
    "django==5.1.6",
    "django-cors-headers==4.7.0",
    "django-debug-toolbar>=5.2.0",
    "django-extensions==3.2.3",
    "django-silk==5.3.2",
    "django-stubs>=5.2.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.5.0",
    "dotenv>=0.9.9",
    "gprof2dot==2024.6.6",
    "gunicorn==23.0.0",
    "hyperlink==21.0.0",
    "idna==3.10",
    "incremental==24.7.2",
    "msgpack==1.1.0",
    "packaging==24.2",
    "pillow==11.1.0",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycodestyle==2.12.1",
    "pycparser==2.22",
    "pyjwt==2.9.0",
    "pyopenssl==25.0.0",
    "redis==5.2.1",
    "service-identity==24.2.0",
    "setuptools==75.8.2",
    "sqlparse==0.5.3",
    "twisted==24.11.0",
    "txaio==23.1.1",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "whitenoise==6.9.0",
    "zope-interface==7.2",
]
